ISSUE #2924 password length and strength moved in yup type
